refactor: moved removal into event/system packages
