+++ engine/src/test/java/org/destinationsol/systems/DestructionSystemTests/RemovalForOptimizationTest.java
@@ -19,7 +19,7 @@ import org.destinationsol.entitysystem.EntitySystemManager;
 import org.destinationsol.game.context.internal.ContextImpl;
 import org.destinationsol.health.components.Health;
 import org.destinationsol.modules.ModuleManager;
-import org.destinationsol.removal.RemovalForOptimizationEvent;
+import org.destinationsol.removal.events.RemovalForOptimizationEvent;
 import org.junit.Before;
 import org.junit.Test;
 import org.terasology.gestalt.entitysystem.component.management.ComponentManager;
